Korekcija kolicine artikla u magacinskoj kartici ^_^

+++ backend/src/daoBean/ArtikalDao.java
@@ -8,7 +8,7 @@ import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDaoBean;
 
 @Stateless
 @Local(ArtikalDaoLocal.class)
-public class ArtikalDao extends GenericDaoBean<Artikal, Long> implements ArtikalDaoLocal{
+public class ArtikalDao extends GenericDaoBean<Artikal, Integer> implements ArtikalDaoLocal{
 
 
 }
