+++ backend/src/daoBean/ArtikalDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Artikal;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface ArtikalDaoLocal extends GenericDao<Artikal, Long>{
+public interface ArtikalDaoLocal extends GenericDao<Artikal, Integer>{
 
 }
